to module torchvision_detection, add pytorch_detection folders
