--- faster_rcnn/train.py
@@ -1,174 +1,0 @@
-import os
-import argparse
-import numpy as np
-import cv2
-import albumentations as A
-from albumentations.pytorch import ToTensorV2
-import torch
-import torchvision
-from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
-from torch.utils.data import DataLoader, Dataset
-from pycocotools.coco import COCO
-from tqdm import tqdm
-
-class CustomDataset(Dataset):
-    def __init__(self, annotation, data_dir, transforms=None):
-        super().__init__()
-        self.data_dir = data_dir
-        self.coco = COCO(annotation)
-        self.predictions = {
-            "images": self.coco.dataset["images"].copy(),
-            "categories": self.coco.dataset["categories"].copy(),
-            "annotations": None
-        }
-        self.transforms = transforms
-
-    def __getitem__(self, index: int):
-        image_id = self.coco.getImgIds(imgIds=index)
-        image_info = self.coco.loadImgs(image_id)[0]
-        
-        image = cv2.imread(os.path.join(self.data_dir, image_info['file_name']))
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB).astype(np.float32)
-        image /= 255.0
-
-        ann_ids = self.coco.getAnnIds(imgIds=image_info['id'])
-        anns = self.coco.loadAnns(ann_ids)
-
-        boxes = np.array([x['bbox'] for x in anns])
-        boxes[:, 2] = boxes[:, 0] + boxes[:, 2]
-        boxes[:, 3] = boxes[:, 1] + boxes[:, 3]
-        
-        labels = np.array([x['category_id']+1 for x in anns]) 
-        labels = torch.as_tensor(labels, dtype=torch.int64)
-        
-        areas = np.array([x['area'] for x in anns])
-        areas = torch.as_tensor(areas, dtype=torch.float32)
-                                
-        is_crowds = np.array([x['iscrowd'] for x in anns])
-        is_crowds = torch.as_tensor(is_crowds, dtype=torch.int64)
-
-        target = {'boxes': boxes, 'labels': labels, 'image_id': torch.tensor([index]), 'area': areas,
-                  'iscrowd': is_crowds}
-
-        if self.transforms:
-            sample = {
-                'image': image,
-                'bboxes': target['boxes'],
-                'labels': labels
-            }
-            sample = self.transforms(**sample)
-            image = sample['image']
-            target['boxes'] = torch.tensor(sample['bboxes'], dtype=torch.float32)
-
-        return image, target, image_id
-    
-    def __len__(self) -> int:
-        return len(self.coco.getImgIds())
-
-def get_train_transform():
-    return A.Compose([
-        A.Resize(1024, 1024),
-        A.Flip(p=0.5),
-        ToTensorV2(p=1.0)
-    ], bbox_params={'format': 'pascal_voc', 'label_fields': ['labels']})
-
-def get_valid_transform():
-    return A.Compose([
-        ToTensorV2(p=1.0)
-    ], bbox_params={'format': 'pascal_voc', 'label_fields': ['labels']})
-
-class Averager:
-    def __init__(self):
-        self.current_total = 0.0
-        self.iterations = 0.0
-
-    def send(self, value):
-        self.current_total += value
-        self.iterations += 1
-
-    @property
-    def value(self):
-        if self.iterations == 0:
-            return 0
-        else:
-            return 1.0 * self.current_total / self.iterations
-
-    def reset(self):
-        self.current_total = 0.0
-        self.iterations = 0.0
-
-def collate_fn(batch):
-    return tuple(zip(*batch))
-
-def train_fn(num_epochs, train_data_loader, optimizer, model, device):
-    best_loss = 1000
-    loss_hist = Averager()
-    for epoch in range(num_epochs):
-        loss_hist.reset()
-
-        for images, targets, image_ids in tqdm(train_data_loader):
-            images = list(image.float().to(device) for image in images)
-            targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
-
-            loss_dict = model(images, targets)
-            losses = sum(loss for loss in loss_dict.values())
-            loss_value = losses.item()
-
-            loss_hist.send(loss_value)
-
-            optimizer.zero_grad()
-            losses.backward()
-            optimizer.step()
-
-        print(f"Epoch #{epoch+1} loss: {loss_hist.value}")
-        if loss_hist.value < best_loss:
-            save_path = './checkpoints/faster_rcnn_torchvision_checkpoints.pth'
-            save_dir = os.path.dirname(save_path)
-            if not os.path.exists(save_dir):
-                os.makedirs(save_dir)
-            
-            torch.save(model.state_dict(), save_path)
-            best_loss = loss_hist.value
-
-def main(args):
-    train_dataset = CustomDataset(args.annotation, args.data_dir, get_train_transform()) 
-    train_data_loader = DataLoader(
-        train_dataset,
-        batch_size=args.batch_size,
-        shuffle=False,
-        num_workers=args.num_workers,
-        collate_fn=collate_fn
-    )
-    device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
-    print(f"Using device: {device}")
-    
-    model = torchvision.models.detection.fasterrcnn_resnet50_fpn(pretrained=True)
-    num_classes = args.num_classes
-    
-    in_features = model.roi_heads.box_predictor.cls_score.in_features
-    model.roi_heads.box_predictor = FastRCNNPredictor(in_features, num_classes)
-    model.to(device)
-    params = [p for p in model.parameters() if p.requires_grad]
-    
-    optimizer = torch.optim.SGD(params, lr=args.lr, momentum=args.momentum, weight_decay=args.weight_decay)
-    
-    train_fn(args.num_epochs, train_data_loader, optimizer, model, device)
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description="Faster R-CNN Training")
-    # 데이터 경로
-    parser.add_argument('--annotation', type=str, default='../dataset/train.json', help='path to annotation file')
-    parser.add_argument('--data_dir', type=str, default='../dataset', help='path to data directory')
-
-    # 학습관련
-    parser.add_argument('--batch_size', type=int, default=16, help='batch size for training')
-    parser.add_argument('--num_workers', type=int, default=8, help='number of worker threads for dataloader')
-    parser.add_argument('--num_classes', type=int, default=11, help='number of classes (including background)')
-    parser.add_argument('--lr', type=float, default=0.005, help='learning rate')
-    parser.add_argument('--momentum', type=float, default=0.9, help='momentum for SGD optimizer')
-    parser.add_argument('--weight_decay', type=float, default=0.0005, help='weight decay for SGD optimizer')
-    parser.add_argument('--num_epochs', type=int, default=12, help='number of epochs to train')
-    
-    args = parser.parse_args()
-    
-    main(args)
