+++ pytorch_detection/src/trainer.py
@@ -1,0 +1,33 @@
+import os
+import torch
+from tqdm import tqdm
+from src.utils import Averager
+
+def train_fn(num_epochs, train_data_loader, optimizer, model, device, save_path):
+    best_loss = 1000
+    loss_hist = Averager()
+    for epoch in range(num_epochs):
+        loss_hist.reset()
+
+        for images, targets, image_ids in tqdm(train_data_loader):
+            images = list(image.float().to(device) for image in images)
+            targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
+
+            loss_dict = model(images, targets)
+            losses = sum(loss for loss in loss_dict.values())
+            loss_value = losses.item()
+
+            loss_hist.send(loss_value)
+
+            optimizer.zero_grad()
+            losses.backward()
+            optimizer.step()
+
+        print(f"Epoch #{epoch+1} loss: {loss_hist.value}")
+        if loss_hist.value < best_loss:
+            save_dir = os.path.dirname(save_path)
+            if not os.path.exists(save_dir):
+                os.makedirs(save_dir)
+            
+            torch.save(model.state_dict(), save_path)
+            best_loss = loss_hist.value
